a26: avoid narrowing conversion in default member initializer
MSVC applies the same narrowing rules to these as it does to list
initializers, and while I don't know if that's justified by the standard
it's easy to avoid narrowing conversions. In this case, basically every
other core already defines System::frequency as a double so the a26 core
was the odd one out.

+++ ares/a26/system/system.hpp
@@ -20,7 +20,7 @@ struct System {
 
   auto name() const -> string { return information.name; }
   auto region() const -> Region { return information.region; }
-  auto frequency() const -> u32 { return information.frequency; }
+  auto frequency() const -> f64 { return information.frequency; }
 
   //system.cpp
   auto game() -> string;
@@ -39,7 +39,7 @@ private:
   struct Information {
     string name = "Atari 2600";
     Region region = Region::NTSC;
-    u32 frequency = Constants::Colorburst::NTSC;
+    f64 frequency = Constants::Colorburst::NTSC;
   } information;
 
   //serialization.cpp
